lag: Sort combined lag file
The combined files are sorted by date, but I'm not sure about the
order within a day because it follows the order in the game logs,
which I'm also not sure about.  I need to touch up some rows, so
having a known game order within a given day will allow me to output
the new file in the same order, helping me check the new file against
the old with diff.