count_wins: Make fewer assumptions
This function hard codes things like the column name because it's just
a convenience function for an operation that I was repeating on the
gamelog data frames.  Still, there's no need to assume that the
away_team and home_team column neighbor each other or that team_type
contains only valid values.