guix: r-rstan: Propagate packages necessary for compilation
I'm not quite sure where to draw the line when I submit r-stan to Guix
proper.  For example, the bash package is needed so that R can find a
shell for its system call in a pure environment.  RStan needs the
system call to compile the R model, but it's not RStan specific.  In
any case, for now it seems clearest to bundle these with the RStan
package definition rather than scattering them in the manifest.