Update rule to extract person-ids.csv
At the time the rule was created, the heading of this table began with
"LAST,FIRST" (see the Wayback Machine [*]).  The current format is now
"ID,Last,First,Player debut,Mgr debut,Coach debut,Ump debut".

I suppose this highlights one of the benefits of using
git-annex/DataLad to store the data files.

[*] http://web.archive.org/web/20170617153245/http://www.retrosheet.org/retroID.htm