Snakefile: Remove lag_download_song2017how_supp_table_
In 24c1eeb (Add PNAS supplmental files via 'git annex addurl',
2018-03-17), I changed this rule to use 'git annex get'.  But there
are two issues with the change: 1) I should have written "{output}"
not "{input}" in the shell command, and 2) even if I had done that
correctly, Snakemake doesn't play nicely with 'git annex get' because
it removes the dead symlink before running the command, but git-annex
needs the symlink to be there.

Instead, just document that those files need to be downloaded if you
want to build the site.  This puts these inputs on the same level as
the inputs/gamelogs submodule in that they need to be set up outside
of Snakemake.