Add a few output files to annex
Snakemake and git-annex don't work that nicely together.  Snakemake
let's you automatically regenerate all the needed dependencies, but it
will abort when it sees locked annex files.  Ignoring the output files
makes rerunning Snakemake easier, but then you can't confirm that the
regenerated files are the same without keeping some sort of
checksum.  (Most of the output files here should have the same
checksum, except for the RDS files for the Stan fits, which have a
timestamp.)

At this point, I'll just store a few of the output files in the annex.
I'm not exposing these files anywhere, so people can't `git annex get`
them, but Snakemake will happily generate a file over a link for a
file that isn't present.  This will make it easy to verify that the
content hasn't changed, and, if a user doesn't want to bother with git
annex, they can just remove the links locally.